Creating the load balancer through the UI:
Lets navigate to user stacks from the menu by going to `Stacks` > `User`. In the `rancher-laravel-demo` stack, we should see an `Add Service` button; lets click the arrow and select `Add Load Balancer`.

First, for the scale pick `Always run one instance of this container on every host.` Next, name the service something
like `laravel-demo-load-balancer` and describe.

The main setup for the load balancer is defining port rules and then mapping them to other services. Our demo is simple, so we are only going to pick `public` `HTTP` on port `80`, and target our `rancher-laravel-demo` service on port `2015`. We leave the `Request Host` input empty, but generally, we would provide our application's DNS hostname here, so Rancher knows how to route it correctly when we have multiple hosts. We leave `path` empty too.

Once we create load balancer service, we should see it activating from the User stacks screen, which might take a few minutes to finish. We can click on the load balancer service to get the host IP address on the `Ports` tab.

If defining a hostname for the load balancer rule, we would grab the IP address from the ports tab and add an A record pointing to this IP. If we click the IP address link from the `Ports` tab in the load balancer service we should see the default Laravel homepage.

The Rancher UI provides a way to copy what we've done in the UI to an equivalent YAML config:

Navigate to `Stacks` > `User` and then click on the `rancher-laravel-demo` stack. The main stack page listing the load balancer and service has a document icon on the top right `Compose YAML`, click it to see the files.

That's what we have done in here: copy > paste.

Test if everything still works well:
`./deploy.sh 06E47AB0BB5DCEEDA0FD nVEntFce51Tdy4ouuesAKA4U6MsUEW2ACQQ3b3fJ http://167.99.46.128:8080/v1/projects/1a5 prod`